# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HabeshaShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USERS
...
  /artists:
    get:
      operationId: listArtists
      tags: [artists]
      summary: List all artists
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of artists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistListResponse"
...
  /artists/{id}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get a specific artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"

  # TRACKS
...
  /tracks/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get a specific track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
...
  /tracks/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a track
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # DISCUSSIONS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, artist, admin]
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        createdAt:
          type: string
...
    Artist:
      type: object
      required: [id, username, displayName, genre, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        genre:
          type: string
        followersCount:
          type: integer
        tracksCount:
          type: integer
        isVerified:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
    ArtistListResponse:
      type: object
      required: [artists, total, page, totalPages]
      properties:
        artists:
          type: array
          items:
            $ref: "#/components/schemas/Artist"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer

    # TRACK
    Track:
      type: object
      required: [id, title, artistId, artistName, genre, youtubeId, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        artistId:
          type: string
        artistName:
          type: string
        artistAvatar:
          type: ["string", "null"]
        genre:
          type: string
        youtubeId:
          type: string
        thumbnail:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        isTrending:
          type: boolean
        createdAt:
          type: string
...
    Comment:
      type: object
      required: [id, trackId, userId, username, text, createdAt]
      properties:
        id:
          type: string
        trackId:
          type: string
        userId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
        text:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500

    # DISCUSSIONS
...